Form: Add ability to select which players to use
Added a multi-select input to choose from the available players. This is
ugly and not in its final form, but the functionality is in place now.
Currently, selecting only one player is treated as though you only want
one player to play at that time, and other players are silenced.